requirements.txt
ipython
numpy
scipy
Pillow
click
flask
